scientifc_agent/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-google-genai==0.6.15
langgraph>=0.0.21
langsmith>=0.0.83
pdfplumber>=0.10.3
python-dotenv>=1.0.0
urllib3>=2.0.0 